Fixed MSVC warnings

+++ tests/tests.cc
@@ -251,7 +251,7 @@ class Tests {
 
     std::string data = "StatObject()";
     std::stringstream ss(data);
-    minio::s3::PutObjectArgs args(ss, data.length(), 0);
+    minio::s3::PutObjectArgs args(ss, static_cast<long>(data.length()), 0);
     args.bucket = bucket_name_;
     args.object = object_name;
     minio::s3::PutObjectResponse resp = client_.PutObject(args);
@@ -284,7 +284,7 @@ class Tests {
     std::string object_name = RandObjectName();
     std::string data = "RemoveObject()";
     std::stringstream ss(data);
-    minio::s3::PutObjectArgs args(ss, data.length(), 0);
+    minio::s3::PutObjectArgs args(ss, static_cast<long>(data.length()), 0);
     args.bucket = bucket_name_;
     args.object = object_name;
     minio::s3::PutObjectResponse resp = client_.PutObject(args);
@@ -301,7 +301,7 @@ class Tests {
 
     std::string data = "DownloadObject()";
     std::stringstream ss(data);
-    minio::s3::PutObjectArgs args(ss, data.length(), 0);
+    minio::s3::PutObjectArgs args(ss, static_cast<long>(data.length()), 0);
     args.bucket = bucket_name_;
     args.object = object_name;
     minio::s3::PutObjectResponse resp = client_.PutObject(args);
@@ -347,7 +347,7 @@ class Tests {
 
     std::string data = "GetObject()";
     std::stringstream ss(data);
-    minio::s3::PutObjectArgs args(ss, data.length(), 0);
+    minio::s3::PutObjectArgs args(ss, static_cast<long>(data.length()), 0);
     args.bucket = bucket_name_;
     args.object = object_name;
     minio::s3::PutObjectResponse resp = client_.PutObject(args);
@@ -436,7 +436,7 @@ class Tests {
       std::string object_name = RandObjectName();
       std::string data = "PutObject()";
       std::stringstream ss(data);
-      minio::s3::PutObjectArgs args(ss, data.length(), 0);
+      minio::s3::PutObjectArgs args(ss, static_cast<long>(data.length()), 0); // PWTODO: PutObjectArgs should accept size_t instead of long
       args.bucket = bucket_name_;
       args.object = object_name;
       minio::s3::PutObjectResponse resp = client_.PutObject(args);
@@ -450,7 +450,7 @@ class Tests {
       std::string object_name = RandObjectName();
       size_t size = 13930573;
       RandCharStream stream(size);
-      minio::s3::PutObjectArgs args(stream, size, 0);
+      minio::s3::PutObjectArgs args(stream, static_cast<long>(size), 0);
       args.bucket = bucket_name_;
       args.object = object_name;
       minio::s3::PutObjectResponse resp = client_.PutObject(args);
@@ -469,7 +469,7 @@ class Tests {
     std::string src_object_name = RandObjectName();
     std::string data = "CopyObject()";
     std::stringstream ss(data);
-    minio::s3::PutObjectArgs args(ss, data.length(), 0);
+    minio::s3::PutObjectArgs args(ss, static_cast<long>(data.length()), 0);
     args.bucket = bucket_name_;
     args.object = src_object_name;
     minio::s3::PutObjectResponse resp = client_.PutObject(args);
@@ -554,7 +554,7 @@ class Tests {
         "1996,Jeep,Grand Cherokee,\"MUST SELL!\n"
         "air, moon roof, loaded\",4799.00\n";
     std::stringstream ss("Year,Make,Model,Description,Price\n" + data);
-    minio::s3::PutObjectArgs args(ss, ss.str().length(), 0);
+    minio::s3::PutObjectArgs args(ss, static_cast<long>(ss.str().length()), 0);
     args.bucket = bucket_name_;
     args.object = object_name;
     minio::s3::PutObjectResponse resp = client_.PutObject(args);
@@ -628,7 +628,7 @@ class Tests {
     try {
       std::string data = "ListenBucketNotification()";
       std::stringstream ss(data);
-      minio::s3::PutObjectArgs args(ss, data.length(), 0);
+      minio::s3::PutObjectArgs args(ss, static_cast<long>(data.length()), 0);
       args.bucket = bucket_name_;
       args.object = object_name;
       minio::s3::PutObjectResponse resp = client_.PutObject(args);
